Fix mypy/pylint CI: add @override, create entity.py, fix type annotations
- Create entity.py with Control4Entity and Control4CoordinatorEntity
  (fixes pylint C7411: entity classes must live in the entity module)
- Add @override to all overriding methods in climate.py (~15 methods)
  and media_player.py (~25 methods); remove now-unused # type: ignore[override]
- Remove unused # type: ignore[misc] from Control4Cover and Control4GarageCover
- Fix device_area: str | None in Control4Climate and Control4CoordinatorEntity
  (mypy arg-type errors when passing item.get("roomName") which can be None)
- Update all platform imports to use .entity directly
